requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
pymongo[srv]==3.12
python-dotenv
PyYAML